move config related files
