+++ app/config/getConfig.ts
