+++ package.json
@@ -79,9 +79,9 @@
     "webpack-merge": "^4.1.1"
   },
   "scripts": {
-    "build:prod": "(cp -n .env.example .env || true) && NODE_ENV=production webpack --config ./webpack/webpack.config.prod.js",
+    "build:prod": "(cp -n ./app/config/.env.example .env || true) && NODE_ENV=production webpack --config ./webpack/webpack.config.prod.js",
     "ganache": "./scripts/ganache-cli.sh",
-    "start": "(cp -n .env.example .env || true) && NODE_ENV=development webpack-dev-server --config ./webpack/webpack.config.dev.js",
+    "start": "(cp -n ./app/config/.env.example .env || true) && NODE_ENV=development webpack-dev-server --config ./webpack/webpack.config.dev.js",
     "format": "prettier --list-different '**/*.{ts,tsx,js,json,md, gql}'",
     "format:fix": "prettier --write '**/*.{ts,tsx,js,json,md,gql}'",
     "tslint": "tslint --project . --format stylish",
